eat: Add script to update Cargo version and modify Tauri config
- New script `update_cargo_version.py` added to update the version in Cargo.toml.
- Refactored `update_tauri_version.py` to modify the version in `tauri.conf.json` instead of Cargo.toml.
- Updated `update_version.sh` to call the new `update_cargo_version.py` and the modified `update_tauri_version.py`.
- Set initial version in `tauri.conf.json` to "0.2.0".
