+++ src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://schema.tauri.app/config/2",
   "productName": "Cuuri",
-  "version": "",
+  "version": "0.2.0",
   "identifier": "com.takanotume24.cuuri",
   "build": {
     "beforeDevCommand": "npm run dev",
